Refactor code into Python package with setup.py Update config handling to use `ovos-config` and match other Diana services Update README.md Add GitHub automation
