+++ docker_overlay/etc/OpenVoiceOS/ovos.conf
@@ -1,0 +1,11 @@
+{
+    "module_overrides": {
+        "diana": {
+            "base_folder": "neon",
+            "config_filename": "diana.yaml"
+        }
+    },
+    "submodule_mappings": {
+        "neon_llm_chatgpt": "diana"
+    }
+}
