--- requirements.txt
@@ -1,2 +1,0 @@
-openai
-neon-mq-connector>=0.6.1a9,~=0.6
