+++ requirements/requirements.txt
@@ -1,0 +1,4 @@
+openai~=0.27
+neon-mq-connector>=0.6.1a9,~=0.6
+ovos-utils~=0.0.32
+ovos-config~=0.0.8
